requirements.txt
Django>=5.2.7
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.4.0
gunicorn>=22.0.0
whitenoise>=6.7.0
pillow>=10.4.0
pytesseract>=0.3.13
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
